create encrypt_module and decrypt_module. Require both in Enigma class and ensure that CLI works with special characters
