+++ decrypted.txt
@@ -1,0 +1,1 @@
+cat hair don't care
