+++ message.txt
@@ -1,1 +1,1 @@
-Spicy mami hot Tamale
+cat Hair Don't care
